Overhaul the backup, restore, and upgrade processes

+++ scripts/app-upgrade.sh
@@ -3,16 +3,22 @@ set -e
 
 START_TIME=$(date +%s)
 
+if [ ! -d .git ]; then
+    echo "[E] This script needs to run from the top directory of the repo. Current working directory:"
+    echo "      $(pwd)"
+    exit 1
+fi
+
 echo "=== Shutting down web container. ==============================================="
 docker-compose stop web
 
 # The backup process will fail if the db container is not started.
 
-echo "=== Starting backup container. ================================================="
-docker-compose up -d backup
+echo "=== Starting cron container. ==================================================="
+docker-compose up -d cron
 
 echo "=== Backing up application stack. =============================================="
-docker-compose exec backup app-backup
+docker-compose exec cron app-backup
 
 echo "=== Removing currnet application stack. ========================================"
 docker-compose down
@@ -27,24 +33,21 @@ echo "[I] Upgrading MunkiReport from '$OLD_MUNKIREPORT_VERSION' to '$NEW_MUNKIRE
 sed -i.bak -e "s/^MUNKIREPORT_VERSION=.*/MUNKIREPORT_VERSION=$NEW_MUNKIREPORT_VERSION/g" .env
 
 echo "=== Deleting old images. ======================================================="
-IMAGE_BACKUP=$(docker images ianharrier/munkireport-backup -q)
+IMAGE_CRON=$(docker images ianharrier/munkireport-cron -q)
 IMAGE_WEB=$(docker images ianharrier/munkireport -q)
-docker rmi $IMAGE_BACKUP $IMAGE_WEB
+docker rmi $IMAGE_CRON $IMAGE_WEB
 
 echo "=== Building new images. ======================================================="
-docker-compose build --pull
+docker-compose build --pull --no-cache
 
 echo "=== Pulling updated database image. ============================================"
 docker-compose pull db
-
-echo "=== Starting backup container. ================================================="
-docker-compose up -d backup
 
 echo "=== Restoring application stack to most recent backup. ========================="
 cd backups
 LATEST_BACKUP=$(ls -1tr *.tar.gz 2> /dev/null | tail -n 1)
 cd ..
-docker-compose exec backup app-restore $LATEST_BACKUP
+./scripts/app-restore.sh $LATEST_BACKUP
 
 END_TIME=$(date +%s)
